Add max_time keyword argument
This addresses https://github.com/litl/backoff/issues/43

The core issue is that with `full_jitter`, `max_tries` is no longer a
reliable way to control the total retry time. This adds a similar
`max_time` keyword argument which triggers a giveup after a total
elapsed amount of time rather than a set number of retries.

Further, since full_jitter is actually the default behavior, you
probably generally want to use `max_time` rather than `max_tries` so
this updates the examples in the README to show that.